refactor: 인증 스토어 파일 구조 변경 및 관련 경로 수정
