+++ src/components/layouts/header.tsx
@@ -6,7 +6,7 @@ import { LinkItem } from '@/types/layout.type';
 import HeaderDropdownMenu from './header-dropdown-menu';
 import { authService } from '@/lib/api/auth-service';
 import { useRouter } from 'next/navigation';
-import { useAuthStore } from '@/store/auth/use-auth-store';
+import { useAuthStore } from '@/store/use-auth-store';
 import { SearchBarComboBox } from '../features/search/search-bar-combo-box';
 
 const Header = () => {
